Fix: DeploymentActvity dashboard api

+++ src/main/java/software/wings/beans/stats/DeploymentActivityStatistics.java
@@ -6,18 +6,18 @@ import java.util.Map;
  * Created by anubhaw on 8/15/16.
  */
 public class DeploymentActivityStatistics extends WingsStatistics {
-  Map<Long, Integer> activitiesCountByDay;
+  Map<Long, Long> activitiesCountByDay;
 
-  public DeploymentActivityStatistics(Map<Long, Integer> activitiesCountByDay) {
+  public DeploymentActivityStatistics(Map<Long, Long> activitiesCountByDay) {
     super(StatisticsType.DEPLOYMENT_ACTIVITIES);
     this.activitiesCountByDay = activitiesCountByDay;
   }
 
-  public Map<Long, Integer> getActivitiesCountByDay() {
+  public Map<Long, Long> getActivitiesCountByDay() {
     return activitiesCountByDay;
   }
 
-  public void setActivitiesCountByDay(Map<Long, Integer> activitiesCountByDay) {
+  public void setActivitiesCountByDay(Map<Long, Long> activitiesCountByDay) {
     this.activitiesCountByDay = activitiesCountByDay;
   }
 }
